test_create_ticket passing
