+++ tests/zendesk_automation/__init__.py
@@ -1,0 +1,4 @@
+#
+# This file is subject to the terms and conditions defined in the
+# file 'LICENSE', which is part of this source code package.
+#
